Fix fullscreen and improve CSD code
Fullscreen mode is currently missing window controls if the maximize button is disabled (as GNOME does by default). There will be a blank space on the right of the navbar instead. This changes it to always show the restore button, which I believe was the intended behavior.

On top of this, I have refactored the code to detect the number and placement of window controls, using variables instead. This code is much cleaner, and has a side effect of now live-updating when the button count or position is changed.